# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conduit API",
        "description": "Conduit is a social blogging site (like Medium) built on top of Laravel. This is the API backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Conduit API server"
        }
    ],
    "paths": {
...
        "/articles/{slug}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get comments for an article.",
                "description": "Get the comments for an article. Auth is optional",
                "operationId": "GetArticleComments",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article that you want to get comments for",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MultipleCommentsResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a comment for an article.",
                "description": "Create a comment for an article. Auth is required",
                "operationId": "CreateArticleComment",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article that you want to create a comment for",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Comment you want to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleCommentResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            }
        },
        "/articles/{slug}/comments/{commentId}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete a comment for an article.",
                "description": "Delete a comment for an article. Auth is required",
                "operationId": "DeleteArticleComment",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article that you want to delete a comment for",
                        "required": true
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID of the comment you want to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            }
        },
...
        "/users": {
            "post": {
                "tags": [
                    "User and Authentication"
                ],
                "summary": "Register a new user.",
                "description": "Register a new user",
                "operationId": "CreateUser",
                "requestBody": {
                    "description": "Details of the new user to register",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "tags": [
                    "User and Authentication"
                ],
                "summary": "Existing user login.",
                "description": "Login for existing user",
                "operationId": "Login",
                "requestBody": {
                    "description": "Credentials to use",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User and Authentication"
                ],
                "summary": "Get current user.",
                "description": "Gets the currently logged-in user",
                "operationId": "GetCurrentUser",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            },
            "put": {
                "tags": [
                    "User and Authentication"
                ],
                "summary": "Update current user.",
                "description": "Updated user information for current user",
                "operationId": "UpdateCurrentUser",
                "requestBody": {
                    "description": "User details to update. At least one field is required.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginUser": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginUserRequest": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/LoginUser"
                    }
                },
                "type": "object"
            },
...
            "NewComment": {
                "properties": {
                    "body": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "NewCommentRequest": {
                "properties": {
                    "comment": {
                        "$ref": "#/components/schemas/NewComment"
                    }
                },
                "type": "object"
            },
            "NewUser": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "NewUserRequest": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/NewUser"
                    }
                },
                "type": "object"
            },
...
            "UpdateUser": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/UpdateUser"
                    }
                },
                "type": "object"
            },
...
            "MultipleCommentsResponse": {
                "properties": {
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    }
                },
                "type": "object"
            },
            "Profile": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "following": {
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object"
            },
...
            "Comment": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "body": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "author": {
                        "$ref": "#/components/schemas/Profile"
                    }
                },
                "type": "object"
            },
            "SingleCommentResponse": {
                "properties": {
                    "comment": {
                        "$ref": "#/components/schemas/Comment"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResponse": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}